+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Next
 - Adapters can now know which columns were requested (#345)
 - Python 3.11 officially supported (#334)
 - Fix for error when an adapater can't be loaded (#346)
+- Fix for ``BestIndexObject`` (#350)
+- Fix for empty dataframes (#351)
 
 Version 1.2.0 - 2023-02-17
 ==========================